# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Gearbox API",
        "description": "REST API untuk Sistem Booking & Manajemen Layanan Bengkel Mobil",
        "contact": {
            "email": "admin@gearbox.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/reviews": {
            "get": {
                "tags": [
                    "Reviews"
                ],
                "summary": "Get semua review (public)",
                "operationId": "9cfedb9e60a10482ea2d0017aa8b381d",
                "parameters": [
                    {
                        "name": "service_id",
                        "in": "query",
                        "description": "Filter berdasarkan ID layanan",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Jumlah data per halaman",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List review dengan pagination",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "current_page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "booking_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "service_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "rating": {
                                                        "type": "integer",
                                                        "example": 5
                                                    },
                                                    "comment": {
                                                        "type": "string",
                                                        "example": "Pelayanan sangat memuaskan!"
                                                    },
                                                    "user": {
                                                        "type": "object"
                                                    },
                                                    "service": {
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": 10
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Reviews"
                ],
                "summary": "Buat review untuk booking yang sudah completed",
                "description": "Review hanya bisa dibuat untuk booking dengan status 'completed'. Setiap booking hanya bisa direview 1 kali.",
                "operationId": "f9bf17f06508aa2dd6aecbb9771092da",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "booking_id",
                                    "rating"
                                ],
                                "properties": {
                                    "booking_id": {
                                        "description": "ID booking yang sudah completed",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "rating": {
                                        "description": "Rating 1-5",
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "comment": {
                                        "type": "string",
                                        "example": "Pelayanan sangat memuaskan, mekanik profesional!"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Review berhasil ditambahkan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Review berhasil ditambahkan"
                                        },
                                        "review": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "booking_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "service_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "rating": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "comment": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Booking tidak ditemukan"
                    },
                    "422": {
                        "description": "Booking belum completed atau sudah pernah direview"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/reviews/{id}": {
            "delete": {
                "tags": [
                    "Admin"
                ],
                "summary": "Hapus review (Admin)",
                "operationId": "d65125ba2fe6a85f90d62e762604b2e2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID review",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Review berhasil dihapus",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Review berhasil dihapus"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Unauthorized — bukan admin"
                    },
                    "404": {
                        "description": "Review tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Masukkan token dari response login",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}